Ajouter le libellé d'une modalité de variable de ventilation

Ajouter le libellé d'une modalité de variable de ventilation (issue d'un fichier du RP Insee) à partir du champ contenant le code en format court.

ajout_libelles_varventil_insee(TABLE, VAR, MILLESIME_RP)

Arguments

TABLE
Table en entrée avec un champ contenant les modalités d'une variable de ventilation en format court (liste : cf. details ci-dessous).

VAR
Nom du champ contenant les modalités d'une variable de ventilation en format court (liste : cf. details ci-dessous).

MILLESIME_RP
Millésime du recensement RP de l'Insee de la variable de ventilation en entrée (liste : cf. details ci-dessous).

Value

Renvoie une table avec un champ supplémentaire contenant le libellé de la modalité.

Details

Les millésimes du RP disponibles sont les suivants : 2008, 2013, 2014, 2015.

Les variables de ventilation disponbibles sont issues de la documentation des fichiers détail 'Migrations Résidentielles' et sont les suivants :

Période regroupée d'achèvement de la construction de la maison ou de l'immeuble ('ACHLR')

Âge regroupé de la personne de référence du ménage en 8 classes d'âge ('AGEMEN8')

Âge quinquennal en années révolues ('AGEREVQ')

Ancienneté d'emménagement ('ANEMC')

Catégorie de population condensée ('CATPC')

Catégorie socioprofessionnelle en 8 postes ('CS1')

Catégorie socioprofessionnelle de la personne de référence du ménage en 8 postes ('CSM')

Diplôme le plus élevé ('DIPL_15')

Département de naissance (si né en France) ('DNAI')

Condition d'emploi ('EMPL')

Indicateur du lieu de naissance ('INAI')

Indicateur de nationalité condensé (Français/Étranger) ('INATC')

Indicateur de résidence antérieure au 1er janvier de l'année précédente ('IRAN')

Indicateur urbain du lieu de résidence antérieure au 1er janvier de l'année précédente ('IRANUU')

Lien à la personne de référence du ménage ('LPRM')

Indicateur Métropole ou DOM du lieu de résidence ('METRODOM')

Mode de cohabitation ('MOCO')

Activité économique en 17 postes (NA - A17) ('NA17')

Activité économique regroupée en 5 postes ('NA5')

Nombre de personnes du ménage (regroupé) ('NPERR')

Ancienneté de recherche d'emploi ('RECH')

Sexe ('SEXE')

Statut d'occupation détaillé du logement ('STOCD')

Type d'activité ('TACT')

Type d'activité de la personne de référence du ménage ('TACTM')

Mode de transport principal le plus souvent utilisé pour aller travailler ('TRANS')

Type de construction ('TYPC')

Type de logement ('TYPL')

Type de ménage regroupé (en 9 postes) ('TYPMR')

Examples

# Ajout du champ 'LIB_CS1' contentant le libellé en clair du code CS1 (aka "groupe socioprofessionnel")
indics_mig_DEP_CS1 <-
ajout_libelles_varventil_insee(TABLE = indics_mig_DEP_CS1,
                               VAR ="CS1",
                               MILLESIME_RP = 2015)

#> Error in eval(lhs, parent, parent): object 'indics_mig_DEP_CS1' not found